requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-filter==24.1
markdown==3.5.2
openai==1.12.0
anthropic==0.18.1
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.2.0